api/requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
python-multipart
pydantic
pydantic-settings
python-dotenv
asyncpg
websockets
httpx
